leftovers
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3978 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ dialogs/LayersListDialog.cpp
@@ -412,7 +412,7 @@ int LayerTreeView::GetItemCount() const
         children.push_back(item);
         item = GetNextItem(item, TVGN_NEXT);
     }
-    count = children.size();
+    count = (int)children.size();
     for (HTREEITEM child : children)
     {
         HTREEITEM subItem = GetChild(child);
